+++ gcdt/ramuda_main.py
@@ -16,7 +16,8 @@ from gcdt import utils
 from gcdt.logger import setup_logger
 from gcdt.ramuda_core import list_functions, get_metrics, deploy_lambda, \
     wire, bundle_lambda, unwire, delete_lambda, rollback, ping, info, cleanup_bundle
-from gcdt.utils import read_gcdt_user_config, get_context, get_command
+from gcdt.utils import read_gcdt_user_config, get_context, get_command, \
+    read_gcdt_user_config_value
 from gcdt.monitoring import datadog_notification, datadog_error, \
     datadog_event_detail
 
@@ -88,6 +89,8 @@ def main():
     elif arguments['deploy']:
         are_credentials_still_valid()
         slack_token, slack_channel = get_user_config()
+        fail_deployment_on_unsuccessful_ping = read_gcdt_user_config_value(
+            'ramuda.failDeploymentOnUnsuccessfulPing', False)
         conf = read_lambda_config()
         lambda_name = conf.get('lambda.name')
         lambda_description = conf.get('lambda.description')
@@ -100,16 +103,16 @@ def main():
         subnet_ids = conf.get('lambda.vpc.subnetIds', None)
         security_groups = conf.get('lambda.vpc.securityGroups', None)
         artifact_bucket = conf.get('deployment.artifactBucket', None)
-        fail_deploy_with_unsuccessfull_ping = conf.get(
-            'deployment.failDeployWithUnsuccessfullPing', False)
         exit_code = deploy_lambda(lambda_name, role_arn, handler_filename,
                                   lambda_handler, folders_from_file,
                                   lambda_description, timeout,
                                   memory_size, subnet_ids=subnet_ids,
                                   security_groups=security_groups,
                                   artifact_bucket=artifact_bucket,
-                                  fail_deploy_with_unsuccessfull_ping=
-                                  fail_deploy_with_unsuccessfull_ping)
+                                  fail_deployment_on_unsuccessful_ping=
+                                  fail_deployment_on_unsuccessful_ping,
+                                  slack_token=slack_token,
+                                  slack_channel=slack_channel)
         event = 'ramuda bot: deployed lambda function: %s ' % lambda_name
         datadog_event_detail(context, event)
     elif arguments['delete']: