+++ tests/test_utils.py
@@ -5,7 +5,7 @@ from nose.tools import assert_equal
 from StringIO import StringIO
 from gcdt import utils
 from gcdt.utils import version, __version__, retries, configure, \
-    read_gcdt_user_config, get_context, get_command, read_gcdt_config_value
+    read_gcdt_user_config, get_context, get_command, read_gcdt_user_config_value
 from .helpers import create_tempfile, cleanup_tempfiles
 
 
@@ -128,20 +128,20 @@ def test_read_user_config_comp_mode():
 
 
 def test_read_gcdt_user_config_value(cleanup_tempfiles):
-    tf = create_tempfile('gcdt {\nfailedDeployWithUnsuccessfullPing=true\n}')
+    tf = create_tempfile('ramuda {\nfailDeploymentOnUnsuccessfulPing=true\n}')
     cleanup_tempfiles.append(tf)
 
-    value = read_gcdt_config_value('failedDeployWithUnsuccessfullPing',
-                                   gcdt_file=tf)
+    value = read_gcdt_user_config_value(
+        'ramuda.failDeploymentOnUnsuccessfulPing', gcdt_file=tf)
     assert value is True
 
 
 def test_read_gcdt_user_config_value_default(cleanup_tempfiles):
-    tf = create_tempfile('gcdt {\nfailedDeployWithUnsuccessfullPing=true\n}')
+    tf = create_tempfile('ramuda {\nfailDeploymentOnUnsuccessfulPing=true\n}')
     cleanup_tempfiles.append(tf)
 
-    value = read_gcdt_config_value('thisValueIsNotPresent', default='my_default',
-                                   gcdt_file=tf)
+    value = read_gcdt_user_config_value('ramuda.thisValueIsNotPresent',
+                                        default='my_default', gcdt_file=tf)
     assert value == 'my_default'